requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
numpy
pandas
pdfminer.six
requests
beautifulsoup4
tiktoken
python-dotenv
langchain
pytest
